mv getStatusCodeFromError() to result helpers.
