+++ src/index.ts
@@ -6,11 +6,8 @@ export {
   defaultEndpointsFactory,
   arrayEndpointsFactory,
 } from "./endpoints-factory";
-export {
-  getExamples,
-  getMessageFromError,
-  getStatusCodeFromError,
-} from "./common-helpers";
+export { getExamples, getMessageFromError } from "./common-helpers";
+export { getStatusCodeFromError } from "./result-helpers";
 export { BuiltinLogger } from "./builtin-logger";
 export { Middleware } from "./middleware";
 export {
